Move Audio controls to widget popup
