+++ packages/audio-input/src/AudioGlobalPanel.tsx
@@ -3,6 +3,9 @@ import { HedronEngine } from '@hedron/engine'
 import { AudioInput } from './AudioInput'
 import { ControlGrid, Param } from '@hedron/ui-core'
 import styles from './AudioGlobalPanel.module.css'
+import { AudioDebugPanel } from './AudioDebugPanel'
+import { FreqPreview } from './FreqPreview'
+import { AudioInputSelector } from './AudioInputSelector'
 
 interface AudioGlobalPanelProps {
   engine: HedronEngine
@@ -37,11 +40,14 @@ export const AudioGlobalPanel: React.FC<AudioGlobalPanelProps> = ({ engine }) =>
   return (
     <div className={styles.panel}>
       <h3 className={styles.title}>Audio Global Settings</h3>
+      <AudioInputSelector audioPlugin={audioPlugin} />
+      <FreqPreview audioPlugin={audioPlugin} />
       <ControlGrid className={styles.controlGrid}>
         {globalOptionNodeIds.map((id: string) => (
           <Param key={id} paramId={id} />
         ))}
       </ControlGrid>
+      <AudioDebugPanel audioPlugin={audioPlugin} />
     </div>
   )
 }
