+++ packages/audio-input/src/AudioInputPanel.module.css
@@ -39,12 +39,6 @@
     border-radius: 4px;
 }
 
-.audioStatusText {
-    font-size: 12px;
-    color: #aaa;
-    margin-bottom: 4px;
-}
-
 .audioStatusActive {
     color: #4CAF50;
 }
