+++ packages/audio-input/src/AudioUtils.ts
@@ -12,126 +12,6 @@ export function createAudioContext(): AudioContext {
 }
 
 /**
- * Interface for test tone options
- */
-export interface TestToneOptions {
-  /** Frequency in Hz */
-  frequency: number
-  /** Volume (0-1) */
-  volume: number
-  /** Wave type */
-  type?: OscillatorType
-}
-
-/**
- * Creates and configures an oscillator for test tones
- * @param audioContext The AudioContext to use
- * @param options Configuration options for the tone
- * @returns Object containing the oscillator and gain nodes
- */
-export function createOscillator(
-  audioContext: AudioContext,
-  options: TestToneOptions,
-): {
-  oscillator: OscillatorNode
-  gainNode: GainNode
-} {
-  // Create oscillator
-  const oscillator = audioContext.createOscillator()
-  oscillator.type = options.type || 'sine'
-  oscillator.frequency.value = options.frequency
-
-  // Create gain node for volume control
-  const gainNode = audioContext.createGain()
-  gainNode.gain.value = options.volume
-
-  // Connect nodes
-  oscillator.connect(gainNode)
-  gainNode.connect(audioContext.destination)
-
-  return { oscillator, gainNode }
-}
-
-/**
- * Plays a test tone with the specified parameters
- * @param options Configuration options or frequency (for backward compatibility)
- * @param duration Duration in milliseconds (0 for continuous)
- * @param volume Volume level from 0-1 (only used if first param is a number)
- * @returns Object with audio nodes and a stop function, or a Promise if duration is provided
- */
-export function playTestTone(
-  options: TestToneOptions | number,
-  duration?: number,
-  volume?: number,
-):
-  | Promise<void>
-  | {
-      audioContext: AudioContext
-      oscillator: OscillatorNode
-      gainNode: GainNode
-      stop: () => void
-    } {
-  // Process parameters for backward compatibility
-  const toneOptions: TestToneOptions =
-    typeof options === 'number'
-      ? { frequency: options, volume: volume || 0.5, type: 'sine' }
-      : options
-
-  const toneDuration = typeof options === 'number' ? duration || 1000 : duration || 0
-
-  // Create audio context and oscillator
-  const audioContext = createAudioContext()
-  const { oscillator, gainNode } = createOscillator(audioContext, toneOptions)
-
-  // Log info about the tone
-  console.log(
-    `[AudioTest] Playing ${toneOptions.frequency}Hz test tone${toneDuration > 0 ? ` for ${toneDuration}ms` : ''} at volume ${toneOptions.volume}`,
-  )
-
-  // Create the control object with stop function
-  const toneControl = {
-    audioContext,
-    oscillator,
-    gainNode,
-    stop: () => {
-      if (timeoutId) {
-        window.clearTimeout(timeoutId)
-      }
-      try {
-        oscillator.stop()
-        audioContext.close()
-        console.log('[AudioTest] Test tone stopped')
-      } catch (err) {
-        // Ignore errors that might occur if already stopped
-      }
-    },
-  }
-
-  // Start oscillator
-  oscillator.start()
-
-  // Setup automatic stop if duration is provided
-  let timeoutId: number | undefined
-  if (toneDuration > 0) {
-    // If duration is specified, return a Promise
-    return new Promise<void>((resolve, reject) => {
-      try {
-        timeoutId = window.setTimeout(() => {
-          toneControl.stop()
-          resolve()
-        }, toneDuration)
-      } catch (error) {
-        console.error('[AudioTest] Error in test tone:', error)
-        reject(error)
-      }
-    })
-  }
-
-  // Otherwise return the control object
-  return toneControl
-}
-
-/**
  * Calculates the average level from an array of audio level values
  * @param levels Array of audio level values (0-1)
  * @returns Average level (0-1)
