+++ packages/ui-core/src/components/WidgetStrip/WidgetStrip.module.css
@@ -1,7 +1,8 @@
 .widgetStrip {
     display: flex;
     flex-direction: column;
-    gap: 10px;
+    width: 100%;
+    height: 100%;
 }
 
 .widgetContainer {
@@ -20,6 +21,10 @@
     padding: 8px;
     animation: slideIn 0.2s ease;
     position: relative;
+    width: 100%;
+    flex: 1;
+    overflow: auto;
+    min-height: 0;
 }
 
 @keyframes slideIn {
